requirements.txt
Flask>=3.1.0
Flask-SQLAlchemy>=3.1.0
Flask-WTF>=1.2.0
WTForms>=3.2.0
email-validator>=2.0.0
SQLAlchemy>=2.0.0
psycopg2-binary>=2.9.9
gunicorn>=23.0.0
python-dotenv>=1.1.0
icalendar>=6.0.0
Pillow>=11.0.0
python-dateutil>=2.9.0
nh3>=0.2.0
werkzeug>=3.1.0
flask-cors>=5.0.0
qrcode[pil]>=8.0
holidays>=0.65
tzdata>=2024.1
